[clean; move] managedfly.bukkitをflatにする
